+++ CHANGELOG.md
@@ -8,6 +8,9 @@ TBD
 
 # 0.1.10
 * **SBBHeaderBox:** reduced margin to fit design specs
+* **SBBTextInput:** renamed from SBBTextField
+* **SBBTextInput:** refactored API and component
+* **SBBTextarea:** added component
 
 # 0.1.9